Decision Engine Refactor
* client.py --> added try except for connection refused, OSError
* header.py --> added more imports
* main.py --> renamed from new_main.py
* main_with_menu.py --> refactored menu, added quarantine, renamed options,
* ness_main.py--> adapted get_table to get the information directly from the sectable
* utils.py --> from continous authentication, to clean main.py
* utiils.py --> from mesh 1.5 to exchange table, etc. to clean main.py
