+++ modules/sc-mesh-secure-deployment/src/1_5/features/continuous/utils.py
@@ -1,0 +1,122 @@
+import multiprocessing
+import pandas as pd
+from time import time, sleep
+import random
+import numpy as np
+import sys
+
+from .ca_main import *
+from ..mutual import mutual
+
+sys.path.append('../../..')
+from common import ConnectionMgr, mesh_utils
+
+co = ConnectionMgr.ConnectionMgr()
+MUTUALINT = 'wlan1'
+MESHINT = 'bat0'
+
+client_q = {}
+
+
+async def launchCA(sectable):
+    """
+    this function should start server within localhost and client within the neighbors
+    It should be the trigger in time-bases ex: every X seconds
+    """
+    ID = random.randint(1000, 64000)
+    with contextlib.suppress(OSError):
+        ca = CA(ID)
+    myip = co.get_ip_address(MESHINT)
+    max_count = 3
+    flag_ctr = 0
+    server_proc = ca_server(myip)
+    neigh = mesh_utils.get_arp()
+    ip2_send = list(set(sectable['IP'].tolist() + list(neigh.keys())))
+    for IP in ip2_send:
+        if IP != myip:
+            flag_ctr = ca_client(IP, flag_ctr, max_count, ca)
+    server_proc.terminate()
+    server_proc.join()
+    return client_q
+
+
+# if not mod:
+#     neigh = mesh_utils.get_macs_neighbors()
+#     for ne in neigh:
+#         if ne not in sectable["MAC"].values:  # means that this is neighbor but was authenticated by someone else
+#             ne_ips = mesh_utils.get_neighbors_ip()
+#             for ip in ne_ips:
+#                 if ip not in sectable["IP"].values:
+#                     info = {'ID': '---', 'MAC': ne, 'IP': ip,
+#                             'PubKey_fpr': "___", 'MA_level': -1}
+#                     mut = mutual.Mutual(MUTUALINT)
+#                     mut.update_table(info)
+#         sectable = pd.read_csv('aux/dev.csv')
+#         ind = sectable.index[sectable['MAC'] == ne].tolist()[0]
+#         IP = sectable.loc[ind]["IP"]
+#         if IP != myip:
+#             flag_ctr = ca_client(IP, flag_ctr, max_count, ca)
+# else:  # modular version MAC are not the ones in neighbors but are the bat0
+#     neigh = mesh_utils.get_arp()
+
+
+def ca_server(myip):
+    with contextlib.suppress(OSError):
+        ca = CA(random.randint(1000, 64000))
+    proc = multiprocessing.Process(target=ca.as_server, args=(myip,), daemon=True)
+    proc.start()
+    sleep(random.randint(1, 10))
+    return proc
+
+
+def ca_client(IP, flag_ctr, max_count, ca):
+    print("neighbor IP:", IP)
+    manager = multiprocessing.Manager()
+    return_dict = manager.dict()
+    p = multiprocessing.Process(target=ca.as_client, args=(IP, return_dict), daemon=True)
+    jobs = [p]
+    p.start()
+    for proc in jobs:
+        proc.join()
+    sleep(5)
+    for key in return_dict.keys():
+        count = np.unique(return_dict[key], return_counts=True)
+        if len(count[1]) > 1:
+            final = count[0][0] if count[1][0] > count[1][1] else count[0][1]
+        else:
+            final = count[0][0]
+    try:
+        client_q[IP] = final
+    except UnboundLocalError:
+        print("Connection Refused")
+        pass
+    flag_ctr += 1
+    if flag_ctr == max_count:
+        p.terminate()
+        # break
+    return flag_ctr
+
+
+def update_table_ca(df, result):
+    """
+    function to update the mutual authentication table with the CA result.
+    Note: this function should be called after the CA result is received.
+    Note2: the new table is not being saved only converted to json and sent to the neighbors.
+    """
+    myID = int(list(set(df.loc[df['IP'] == co.get_ip_address(MESHINT), "ID"]))[0])
+    df = df.assign(CA_Result=0)
+    for index, row in df.iterrows():
+        for res in result:
+            for ip in res.keys():
+                if row['IP'] == ip:
+                    IP = row['IP']
+                    if res[ip] in ['pass', 1]:
+                        df.loc[index, 'CA_Result'] = 1
+                    elif res[ip] == 'fail':
+                        df.loc[df['IP'] == IP, 'CA_Result'] = 2
+                    else:
+                        df.loc[df['IP'] == IP, 'CA_Result'] = 3
+                    # df.loc[df['IP'] == IP, 'CA_ts'] = time()
+                    df.loc[df['IP'] == IP, 'CA_server'] = myID
+    df.drop_duplicates(inplace=True)
+    return df
