+++ modules/sc-mesh-secure-deployment/src/1_5/features/ness/ness_main.py
@@ -4,6 +4,8 @@ import pickle
 import os
 import json
 import time
+import numpy as np
+import pandas as pd
 
 file_path = os.path.dirname(__file__)
 
@@ -17,14 +19,6 @@ class NESS:
     def create_status_list(self, sec_list, n):
         return [sec_list[i][len(sec_list[i]) - 1] for i in range(n)]
 
-    def create_good_server_list(self, sec_list, n):
-        out = []
-        for i in range(n):
-            l = len(sec_list[i])
-            if sec_list[i][l - 1] == 1:
-                out.append(i)
-        return out
-
     def mapping(self, nodes):
         mapp = dict(zip(nodes, range(len(nodes))))
         print("Map node labels with ID :", mapp)
@@ -36,8 +30,19 @@ class NESS:
         val_list = list(mapps.values())
         return key_list[position]
 
+    def create_good_server_list(self, sec_list, n):
+        out = []
+        for i in range(n):
+            l = len(sec_list[i])
+            if sec_list[i][l - 1] == 1:
+                out.append(i)
+        return out
+
     def create_servers_flags_list(self, sec_list, n, p):
-        return [sec_list[i][p] for i in range(n)]
+        out = []
+        for i in range(n):
+            out.append(sec_list[i][p])
+        return out
 
     def save_file_status(self, **args):
         aux = {arg: args[arg] for arg in args}
@@ -46,21 +51,34 @@ class NESS:
         with open("last_result.json", "w") as outfile:
             outfile.write(jsonStr)
 
-    def first_table(self, df, laststatus=None):
-        '''
-        Assuming that timestamp is small enough between rows
-        '''
-        nodes = df['IP'].to_list()
-        latest_status_list = laststatus or [1 for _ in range(df.shape[0])]
-        good_server_status_list = list(range(len(nodes)))
-        flags_list = df["CA_Result"].tolist()
-        mapp = self.mapping(df["ID"].tolist())
-        servers_list = list(mapp.values())
-        n = df.shape[0]
-        self.save_file_status(latest_status_list=latest_status_list, good_server_status_list=good_server_status_list, flags_list=flags_list, servers_list=servers_list, mapp=mapp)
-        return latest_status_list, good_server_status_list, flags_list, servers_list, n, mapp
+    def get_table(self, df):
+        flags_map = {65: 1, 131: 2, 132: 2, 133: 2, 134: 2, 194: 3}  # needs to add disconection (-1)
+        mappids = self.mapping(set(df["ID"].tolist()))
+        df2 = df.replace({"CA_server": mappids})
+        # df2 = df2.replace({"ID": mapp})
+        new_list = []
+        last_status = []
+        for i in df2["ID"]:
+            df2.loc[df2['CA_server'].isnull(), "CA_server"] = list(pd.Series(i).map(mappids))[0]
+            table = [list(pd.Series(i).map(mappids))[0]]
+            servers = list(df2.loc[df2['ID'] == i, "CA_server"])
+            table.append(servers)
+            flags = list(df2.loc[df2['ID'] == i, "CA_Result"])
+            table.append(flags)
+            count = np.unique(flags, return_counts=True)
+            if len(count[1]) > 1:
+                final = count[0][0] if count[1][0] > count[1][1] else count[0][1]
+            else:
+                final = count[0][0]
+            table.append(final)
+            new_list.append(table)
+        finaltable = []
+        for l in new_list:
+            if l not in finaltable:
+                finaltable.append(l)
+        return finaltable, mappids
 
-    def adapt_table(self, result):
+    def adapt_table(self, result, mapp):
         if not os.path.isfile('last_result.json'):
             latest_status_list = []
             good_server_status_list = []
@@ -75,19 +93,19 @@ class NESS:
             servers_list = json_object['servers_list']
             mapp = json_object['mapp']
         for node in result:
-            if result[node] == '65':
+            if result[node] == 65:
                 latest_status_list.append(1)
                 good_server_status_list.append(node)
                 flags_list.append(1)
-            if result[node] in ['131', '132', '133', '134']:
+            if result[node] in [131, 132, 133, 134]:
                 latest_status_list.append(2)
                 flags_list.append(2)
-            if result[node] == '194':
+            if result[node] == 194:
                 latest_status_list.append(3)
                 flags_list.append(3)
         n = len(servers_list)
-        self.save_file_status(latest_status_list=latest_status_list, good_server_status_list=good_server_status_list, flags_list=flags_list, servers_list=servers_list, mapp=mapp)
-        return latest_status_list, good_server_status_list, flags_list, servers_list, n, mapp
+        self.save_file_status(latest_status_list=latest_status_list, good_server_status_list=good_server_status_list,
+                              flags_list=flags_list, servers_list=servers_list, mapp=mapp)
 
     def run_decision(self, latest_status_list, good_server_status_list, flags_list, servers_list, n, i):
         self.engine.reset()
@@ -168,35 +186,68 @@ class NESS:
 
         self.run(init_latest_status_list, init_good_server_status_list, init_flags_table, init_servers_table, n)
 
-    def run(self, init_latest_status_list, init_good_server_status_list, init_flags_table, init_servers_table, n):
-        result = {}
-        print("Running Decision on all nodes as the Sec Table is valid")
-        print("Nodes status table: ", init_latest_status_list)
-        print("Known good Servers table: ", init_good_server_status_list)
-        print("All Servers table: ", init_servers_table)
-        print("All Flags table: ", init_flags_table)
-        n_list = [n]
-        for i in range(n):
-            #
-            # Marshalling layer:
-            # index of node to check
-            #
-            i_list = [i]
-            #
-            # Preparing tuples for the query into Pyke engine
-            #
-            latest_status_list = tuple(init_latest_status_list)
-            good_server_status_list = tuple(init_good_server_status_list)
-            # flags_list = tuple(init_flags_table[i])
-            # servers_list = tuple(init_servers_table[i])
-            flags_list = tuple(init_flags_table)
-            servers_list = tuple(init_servers_table)
-            nt = tuple(n_list)
-            it = tuple(i_list)
+    def ness_result_to_table(self, df, ness_result, mapp):
+        df = df.assign(Ness_Result=0)
+        for res in ness_result:
+            ind = self.remapping(mapp, res)
+            df.loc[df['ID'] == ind, 'Ness_Result'] = int(ness_result[res])
+        df.drop_duplicates(inplace=True)
+        return df
 
-            act_code, mnode = self.run_decision(latest_status_list, good_server_status_list, flags_list, servers_list,
-                                                nt, it)
-            print("\nAction Code issued is ", act_code)
-            print("\nfor node ", mnode)
-            result[mnode] = act_code
+    def run_all(self, output, latest_status=False):
+        result = {}
+        T_struct = output
+        n = len(T_struct)
+        n_list = [n]
+        sec_table_valid = 1
+        init_latest_status_list = self.create_status_list(T_struct, n)
+        init_latest_status_list1 = []
+        init_latest_status_list1.append(init_latest_status_list)
+        init_good_server_status_list = self.create_good_server_list(T_struct, n)
+        p = 1
+        init_servers_table = self.create_servers_flags_list(T_struct, n, p)
+        p = 2
+        init_flags_table = self.create_servers_flags_list(T_struct, n, p)
+        #
+        # Marshalling to prepare dashboard and Decision Function call if table valid
+        #
+        if sec_table_valid == 1:
+            print("Running Decision on all nodes as the Sec Table is valid")
+            print("Nodes status table: ", init_latest_status_list)
+            print("Known good Servers table: ", init_good_server_status_list)
+            print("All Servers table: ", init_servers_table)
+            print("All Flags table: ", init_flags_table)
+            for i in range(n):
+                #
+                # Marshalling layer:
+                # index of node to check
+                #
+                i_list = []
+                i_list.append(i)
+                #
+                # Preparing tuples for the query into Pyke engine
+                #
+                if latest_status:
+                    f = open('last_result.json')
+                    json_object = json.load(f)
+                    latest_status_list = json_object['latest_status_list']
+                    good_server_status_list = json_object['good_server_status_list']
+                    flags_list = json_object['flags_list']
+                    servers_list = json_object['servers_list']
+                    mapp = json_object['mapp']
+                else:
+                    latest_status_list = tuple(init_latest_status_list)
+                    good_server_status_list = tuple(init_good_server_status_list)
+                    flags_list = tuple(init_flags_table[i])
+                    servers_list = tuple(init_servers_table[i])
+                nt = tuple(n_list)
+                it = tuple(i_list)
+
+                act_code, nnode = self.run_decision(latest_status_list, good_server_status_list, flags_list,
+                                                    servers_list, nt,
+                                                    it)
+                print("\nFor node: ", nnode)
+                print("Action Code issued is ", act_code)
+
+                result[nnode] = act_code
         return result
